+++ Sprint04/lib/constants.js
@@ -2,10 +2,8 @@ import dotenv from 'dotenv';
 
 dotenv.config();
 
-const NODE_ENV = process.env.NODE_ENV || 'development';
 const PORT = process.env.PORT || 3000;
+const ACCESS_TOKEN_SECRET = process.env.ACCESS_TOKEN_SECRET;
 
-export {
-  NODE_ENV,
-  PORT,
-};
+
+export {PORT, ACCESS_TOKEN_SECRET};
