--- Sprint04/routes/auth.js
@@ -1,47 +1,0 @@
-import express from 'express';
-import bcrypt from 'bcrypt';
-import prisma from '../lib/prisma.js';
-
-const router = express.Router();
-
-router.post('/auth/register', register);
-
-async function register(req, res) {
-  const { email, nickname, password, image } = req.body;
-
-  if (!email || !nickname || !password || !image) {
-    return res.status(400).json({ message: '모든 필드를 입력해주세요.' });
-  }
-
-  try {
-    const existingUser = await prisma.user.findFirst({
-      where: {
-        OR: [{ email }, { nickname }],
-      },
-    });
-
-    if (existingUser) {
-      return res.status(409).json({ message: '이미 존재하는 이메일 또는 닉네임입니다.' });
-    }
-
-    const salt = await bcrypt.genSalt(10);
-    const hashedPassword = await bcrypt.hash(password, salt);
-    const user = await prisma.user.create({
-      data: {
-        email,
-        nickname,
-        image,
-        password: hashedPassword,
-      },
-    });
-
-    const { password: _, ...userWithoutPassword } = user;
-
-    res.status(201).json(userWithoutPassword);
-  } catch (error) {
-    console.error('회원가입 에러:', error);
-    res.status(500).json({ message: '서버 에러. 잠시 후 다시 시도해주세요.' });
-  }
-}
-
-export default router;
